requirements.txt
streamlit==1.35.0
plotly==5.21.0
psycopg2-binary==2.9.9
pandas==2.2.2
python-dotenv==1.0.1